[scanner] fix: split oversized mcp hook tests (group B)
Split 4 oversized test files in src/hooks/mcp/__tests__ (each over
500 lines) into sibling files by logical describe block, per issue
#22010:

- workloads.core.test.ts (928) -> workloads.core.test.ts (492),
  workloads.core-2.test.ts (383), workloads.core-3.test.ts (367)
- shared-cache-fetch.test.ts (925) -> shared-cache-fetch.test.ts (349),
  shared-cache-fetch-2.test.ts (448), shared-cache-fetch-3.test.ts (438)
- shared-pure-funcs.test.ts (922) -> shared-pure-funcs.test.ts (447),
  shared-pure-funcs-2.test.ts (483), shared-pure-funcs-3.test.ts (304)

helm-core.test.ts was already below 500 lines (441) so it was left
unchanged.

All 181 existing test cases are preserved with identical behavior;
mocks and fixtures are duplicated verbatim into each sibling file.
