+++ web/src/hooks/mcp/__tests__/shared-cache-fetch-2.test.ts
@@ -1,0 +1,448 @@
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
+import type { ClusterInfo, ClusterHealth } from '../types'
+
+// ---------------------------------------------------------------------------
+// Constants used in tests (mirror source values to avoid magic numbers)
+// ---------------------------------------------------------------------------
+const OFFLINE_THRESHOLD_MS = 5 * 60_000 // 5 minutes — same as OFFLINE_THRESHOLD_MS in shared.ts
+const AUTO_GENERATED_NAME_LENGTH_THRESHOLD = 50 // same as in shared.ts
+const CLUSTER_NOTIFY_DEBOUNCE_MS = 50 // same debounce delay in shared.ts
+const DEFAULT_MAX_RETRIES = 2 // fetchWithRetry default
+const DEFAULT_INITIAL_BACKOFF_MS = 500 // fetchWithRetry default
+
+// ---------------------------------------------------------------------------
+// Hoisted mocks
+// ---------------------------------------------------------------------------
+const mockIsDemoMode = vi.hoisted(() => vi.fn(() => false))
+const mockIsDemoToken = vi.hoisted(() => vi.fn(() => false))
+const mockIsNetlifyDeployment = vi.hoisted(() => ({ value: false }))
+const mockSubscribeDemoMode = vi.hoisted(() => vi.fn())
+const mockIsBackendUnavailable = vi.hoisted(() => vi.fn(() => false))
+const mockReportAgentDataError = vi.hoisted(() => vi.fn())
+const mockReportAgentDataSuccess = vi.hoisted(() => vi.fn())
+const mockIsAgentUnavailable = vi.hoisted(() => vi.fn(() => true))
+const mockGetAgentClusterCount = vi.hoisted(() => vi.fn(() => 0))
+const mockRegisterCacheReset = vi.hoisted(() => vi.fn())
+const mockTriggerAllRefetches = vi.hoisted(() => vi.fn())
+const mockResetFailuresForCluster = vi.hoisted(() => vi.fn())
+const mockResetAllCacheFailures = vi.hoisted(() => vi.fn())
+const mockKubectlProxyExec = vi.hoisted(() => vi.fn())
+const mockApiGet = vi.hoisted(() => vi.fn())
+
+vi.mock('../mcp/shared', () => ({
+  agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
+  clusterCacheRef: { clusters: [] },
+  REFRESH_INTERVAL_MS: 120_000,
+  CLUSTER_POLL_INTERVAL_MS: 60_000,
+}))
+
+vi.mock('../../../lib/api', () => ({
+  api: { get: mockApiGet },
+  isBackendUnavailable: mockIsBackendUnavailable,
+}))
+
+vi.mock('../../../lib/demoMode', () => ({
+  isDemoMode: mockIsDemoMode,
+  isDemoToken: mockIsDemoToken,
+  get isNetlifyDeployment() {
+    return mockIsNetlifyDeployment.value
+  },
+  subscribeDemoMode: mockSubscribeDemoMode,
+}))
+
+vi.mock('../../useLocalAgent', () => ({
+  reportAgentDataError: mockReportAgentDataError,
+  reportAgentDataSuccess: mockReportAgentDataSuccess,
+  isAgentUnavailable: mockIsAgentUnavailable,
+  getAgentClusterCount: mockGetAgentClusterCount,
+}))
+
+vi.mock('../../../lib/modeTransition', () => ({
+  registerCacheReset: mockRegisterCacheReset,
+  triggerAllRefetches: mockTriggerAllRefetches,
+}))
+
+vi.mock('../../../lib/cache', () => ({
+  resetFailuresForCluster: mockResetFailuresForCluster,
+  resetAllCacheFailures: mockResetAllCacheFailures,
+  createCachedHook: vi.fn((_config: unknown) => () => ({})),
+}))
+
+vi.mock('../../../lib/kubectlProxy', () => ({
+  kubectlProxy: { exec: mockKubectlProxyExec },
+}))
+
+vi.mock('../../../lib/constants', async () => {
+  const actual = await vi.importActual<typeof import('../../../lib/constants')>('../../../lib/constants')
+  return {
+    ...actual,
+  }
+})
+
+vi.mock('../../../lib/constants/network', async () => {
+  const actual = await vi.importActual<typeof import('../../../lib/constants/network')>('../../../lib/constants/network')
+  return {
+    ...actual,
+  }
+})
+
+// ---------------------------------------------------------------------------
+// Imports (resolved after mocks are installed)
+// ---------------------------------------------------------------------------
+import {
+  // Constants
+  REFRESH_INTERVAL_MS,
+  CLUSTER_POLL_INTERVAL_MS,
+  GPU_POLL_INTERVAL_MS,
+  CACHE_TTL_MS,
+  MIN_REFRESH_INDICATOR_MS,
+  // Pure functions
+  getEffectiveInterval,
+  shareMetricsBetweenSameServerClusters,
+  deduplicateClustersByServer,
+  shouldMarkOffline,
+  recordClusterFailure,
+  clearClusterFailure,
+  clusterDisplayName,
+  fetchWithRetry,
+  _resetAgentTokenState,
+  // Async functions
+  fullFetchClusters,
+  refreshSingleCluster,
+  fetchSingleClusterHealth,
+  connectSharedWebSocket,
+  // State management
+  clusterCache,
+  clusterSubscribers,
+  notifyClusterSubscribers,
+  notifyClusterSubscribersDebounced,
+  updateClusterCache,
+  updateSingleClusterInCache,
+  setInitialFetchStarted,
+  setHealthCheckFailures,
+  getInitialFetchStarted,
+  getHealthCheckFailures,
+  initialFetchStarted,
+  healthCheckFailures,
+  // WebSocket
+  sharedWebSocket,
+  cleanupSharedWebSocket,
+  // Cache ref
+  clusterCacheRef,
+  subscribeClusterCache,
+} from '../shared'
+import { clearAgentToken, setAgentToken } from '../agentFetch'
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+function makeCluster(overrides: Partial<ClusterInfo> = {}): ClusterInfo {
+  return {
+    name: 'test-cluster',
+    context: 'test-context',
+    server: 'https://test.example.com:6443',
+    healthy: true,
+    source: 'kubeconfig',
+    nodeCount: 3,
+    podCount: 20,
+    cpuCores: 8,
+    memoryGB: 32,
+    storageGB: 100,
+    ...overrides,
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Tests
+// ---------------------------------------------------------------------------
+
+describe('localStorage cluster cache persistence', () => {
+  beforeEach(() => {
+    localStorage.clear()
+    clusterSubscribers.clear()
+    updateClusterCache({ clusters: [], isLoading: false })
+  })
+
+  it('saves clusters to localStorage when updateClusterCache is called', () => {
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'persisted' })],
+    })
+    const stored = localStorage.getItem('kubestellar-cluster-cache')
+    expect(stored).not.toBeNull()
+    const parsed = JSON.parse(stored!)
+    expect(parsed.some((c: ClusterInfo) => c.name === 'persisted')).toBe(true)
+  })
+
+  it('filters out clusters with slash in name from localStorage', () => {
+    updateClusterCache({
+      clusters: [
+        makeCluster({ name: 'good-name' }),
+        makeCluster({ name: 'path/with/slash' }),
+      ],
+    })
+    const stored = localStorage.getItem('kubestellar-cluster-cache')
+    const parsed = JSON.parse(stored!)
+    expect(parsed.every((c: ClusterInfo) => !c.name.includes('/'))).toBe(true)
+  })
+
+  it('saves distribution cache to localStorage', () => {
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'dist-test', distribution: 'openshift' })],
+    })
+    const stored = localStorage.getItem('kubestellar-cluster-distributions')
+    expect(stored).not.toBeNull()
+    const parsed = JSON.parse(stored!)
+    expect(parsed['dist-test']).toEqual(expect.objectContaining({ distribution: 'openshift' }))
+  })
+
+  it('applies distribution from localStorage cache to cluster without distribution', () => {
+    // First, save a distribution to cache
+    localStorage.setItem('kubestellar-cluster-distributions', JSON.stringify({
+      'cached-cluster': { distribution: 'eks', namespaces: ['ns1'] }
+    }))
+
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'cached-cluster', distribution: undefined, server: 'https://custom.internal' })],
+    })
+    const c = clusterCache.clusters.find(c => c.name === 'cached-cluster')!
+    expect(c.distribution).toBe('eks')
+    expect(c.namespaces).toEqual(['ns1'])
+  })
+})
+
+// ---------------------------------------------------------------------------
+// mergeWithStoredClusters (private, exercised through updateClusterCache)
+// ---------------------------------------------------------------------------
+describe('mergeWithStoredClusters (via updateClusterCache)', () => {
+  beforeEach(() => {
+    localStorage.clear()
+    clusterSubscribers.clear()
+  })
+
+  it('preserves cached metrics when new cluster data is missing metrics', () => {
+    const CPU_CORES = 16
+    const MEM_GB = 64
+    // Seed localStorage with a cluster that has metrics
+    localStorage.setItem('kubestellar-cluster-cache', JSON.stringify([
+      { name: 'merge-test', context: 'ctx', cpuCores: CPU_CORES, memoryGB: MEM_GB, nodeCount: 5, podCount: 40 }
+    ]))
+
+    // Update with a cluster that has no metrics
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'merge-test', cpuCores: undefined, memoryGB: undefined, nodeCount: undefined, podCount: undefined })],
+    })
+
+    const c = clusterCache.clusters.find(c => c.name === 'merge-test')!
+    expect(c.cpuCores).toBe(CPU_CORES)
+    expect(c.memoryGB).toBe(MEM_GB)
+  })
+
+  it('uses new metrics when they are positive', () => {
+    const OLD_CPU = 8
+    const NEW_CPU = 32
+    localStorage.setItem('kubestellar-cluster-cache', JSON.stringify([
+      { name: 'merge-new', context: 'ctx', cpuCores: OLD_CPU }
+    ]))
+
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'merge-new', cpuCores: NEW_CPU })],
+    })
+
+    const c = clusterCache.clusters.find(c => c.name === 'merge-new')!
+    expect(c.cpuCores).toBe(NEW_CPU)
+  })
+
+  it('preserves health status from cached data when new data is undefined', () => {
+    localStorage.setItem('kubestellar-cluster-cache', JSON.stringify([
+      { name: 'health-merge', context: 'ctx', healthy: true, reachable: true }
+    ]))
+
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'health-merge', healthy: undefined, reachable: undefined })],
+    })
+
+    const c = clusterCache.clusters.find(c => c.name === 'health-merge')!
+    expect(c.healthy).toBe(true)
+    expect(c.reachable).toBe(true)
+  })
+})
+
+// ---------------------------------------------------------------------------
+// fullFetchClusters — demo mode paths
+// ---------------------------------------------------------------------------
+describe('fullFetchClusters', () => {
+  const originalFetch = globalThis.fetch
+
+  beforeEach(() => {
+    localStorage.clear()
+    clusterSubscribers.clear()
+    mockIsDemoMode.mockReturnValue(false)
+    mockIsDemoToken.mockReturnValue(false)
+    mockIsNetlifyDeployment.value = false
+    mockIsAgentUnavailable.mockReturnValue(true)
+    // Reset cache to clean state
+    updateClusterCache({
+      clusters: [],
+      isLoading: true,
+      isRefreshing: false,
+      error: null,
+      consecutiveFailures: 0,
+      isFailed: false,
+      lastUpdated: null,
+      lastRefresh: null,
+    })
+  })
+
+  afterEach(() => {
+    globalThis.fetch = originalFetch
+    vi.restoreAllMocks()
+  })
+
+  it('returns demo clusters when isDemoMode() is true and demo token is set', async () => {
+    // #6243 dropped the unconditional demo-mode short-circuit. Demo data
+    // is now returned only when:
+    //   1) Netlify forced demo, OR
+    //   2) fetchClusterListFromAgent() returns null AND
+    //      isDemoMode() && isDemoToken() are both true.
+    // Tests must mock both flags AND make agent fetch fail so the demo
+    // fallback branch fires.
+    mockIsDemoMode.mockReturnValue(true)
+    mockIsDemoToken.mockReturnValue(true)
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
+    await fullFetchClusters()
+    expect(clusterCache.clusters.length).toBeGreaterThan(0)
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.error).toBeNull()
+    // Demo clusters should include well-known demo names
+    const names = clusterCache.clusters.map(c => c.name)
+    expect(names).toContain('kind-local')
+  })
+
+  it('returns demo clusters on Netlify with demo token', async () => {
+    mockIsNetlifyDeployment.value = true
+    mockIsDemoToken.mockReturnValue(true)
+    localStorage.setItem('token', 'demo-token')
+    await fullFetchClusters()
+    expect(clusterCache.clusters.length).toBeGreaterThan(0)
+    expect(clusterCache.isLoading).toBe(false)
+  })
+
+  it('falls back gracefully on fetch error (no blocking error)', async () => {
+    // Agent unavailable + no token = should finish loading
+    localStorage.removeItem('token')
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('network'))
+    mockApiGet.mockRejectedValue(new Error('network'))
+    await fullFetchClusters()
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.error).toBeNull() // Never sets error
+  })
+
+  it('fetches from backend API when agent is unavailable and token exists', async () => {
+    localStorage.setItem('token', 'real-token')
+    const BACKEND_CLUSTERS = [makeCluster({ name: 'backend-cluster' })]
+    mockApiGet.mockResolvedValue({ data: { clusters: BACKEND_CLUSTERS } })
+    // Agent returns null
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
+
+    await fullFetchClusters()
+
+    // The backend API fallback now uses the same-origin /api/mcp/clusters endpoint
+    // which works regardless of agent backend (kc-agent, kagenti, kagent). (#9535)
+    expect(mockApiGet).toHaveBeenCalledWith('/api/mcp/clusters')
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.clusters.some(c => c.name === 'backend-cluster')).toBe(true)
+  })
+
+  it('fetches from backend API for cookie-only sessions when agent is unavailable', async () => {
+    localStorage.setItem('kc-has-session', 'true')
+    const BACKEND_CLUSTERS = [makeCluster({ name: 'cookie-session-cluster' })]
+    mockApiGet.mockResolvedValue({ data: { clusters: BACKEND_CLUSTERS } })
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
+
+    await fullFetchClusters()
+
+    expect(mockApiGet).toHaveBeenCalledWith('/api/mcp/clusters')
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.clusters.some(c => c.name === 'cookie-session-cluster')).toBe(true)
+  })
+
+  it('routes cluster fetch through backend API when kagenti backend is preferred (#9535)', async () => {
+    localStorage.setItem('kc_agent_backend_preference', 'kagenti')
+    const KAGENTI_CLUSTERS = [makeCluster({ name: 'kagenti-cluster' })]
+    mockApiGet.mockResolvedValue({ data: { clusters: KAGENTI_CLUSTERS } })
+    // globalThis.fetch should NOT be called — kagenti path uses api.get
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('should not be called'))
+
+    await fullFetchClusters()
+
+    expect(mockApiGet).toHaveBeenCalledWith('/api/mcp/clusters')
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.clusters.some(c => c.name === 'kagenti-cluster')).toBe(true)
+  })
+
+  it('skips backend when no auth token or cookie session marker exists', async () => {
+    // The previous test may have set auth state; clear it all.
+    localStorage.clear()
+    mockApiGet.mockClear()
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
+
+    await fullFetchClusters()
+
+    expect(mockApiGet).not.toHaveBeenCalled()
+    expect(clusterCache.isLoading).toBe(false)
+  })
+
+  it('deduplicates concurrent calls (only one runs at a time)', async () => {
+    mockIsDemoMode.mockReturnValue(true)
+    const p1 = fullFetchClusters()
+    const p2 = fullFetchClusters() // Should be a no-op
+    await Promise.all([p1, p2])
+    // Both resolve without error
+    expect(clusterCache.isLoading).toBe(false)
+  })
+
+  it('does not fall back to demo clusters on backend API error when demo mode is disabled', async () => {
+    localStorage.setItem('token', 'real-token')
+    // Agent returns null
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent down'))
+    // Backend API also throws
+    mockApiGet.mockRejectedValue(new Error('backend unavailable'))
+
+    await fullFetchClusters()
+
+    expect(clusterCache.isLoading).toBe(false)
+    expect(clusterCache.error).toBe('Cluster data unavailable')
+    expect(clusterCache.clusters).toEqual([])
+    expect(clusterCache.consecutiveFailures).toBeGreaterThan(0)
+  })
+
+  it('on Netlify with real token, skips early return and tries fetch', async () => {
+    mockIsNetlifyDeployment.value = true
+    localStorage.setItem('token', 'real-user-token')
+    // Agent will fail (Netlify), backend should be tried
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent'))
+    mockApiGet.mockResolvedValue({ data: { clusters: [makeCluster({ name: 'netlify-real' })] } })
+
+    await fullFetchClusters()
+
+    expect(clusterCache.clusters.some(c => c.name === 'netlify-real')).toBe(true)
+  })
+
+  it('preserves existing clusters on fetch error when cache has data', async () => {
+    // Seed some initial clusters
+    updateClusterCache({
+      clusters: [makeCluster({ name: 'existing' })],
+      isLoading: false,
+    })
+
+    localStorage.setItem('token', 'real-token')
+    globalThis.fetch = vi.fn().mockRejectedValue(new Error('agent'))
+    mockApiGet.mockRejectedValue(new Error('backend'))
+
+    await fullFetchClusters()
+
+    // Should preserve existing clusters, not replace with demo
+    expect(clusterCache.clusters.some(c => c.name === 'existing')).toBe(true)
+  })
+})
+
