+++ web/src/hooks/mcp/__tests__/workloads.core.test.ts
@@ -490,439 +490,3 @@ describe('useDeploymentIssues', () => {
   })
 })
 
-// ===========================================================================
-// useDeployments
-// ===========================================================================
-
-describe('useDeployments', () => {
-  it('returns initial loading state with empty deployments array', () => {
-    // Block all fetch paths to keep hook in loading state
-    mockFetchSSE.mockReturnValue(new Promise(() => {}))
-    globalThis.fetch = vi.fn().mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useDeployments())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.deployments).toEqual([])
-  })
-
-  it('returns demo deployments when demo mode is active', async () => {
-    mockIsDemoMode.mockReturnValue(true)
-    mockUseDemoMode.mockReturnValue(true)
-
-    const { result } = renderHook(() => useDeployments())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.deployments.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns deployments from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeDeployments = [
-      { name: 'api', namespace: 'prod', status: 'running', replicas: 3, readyReplicas: 3, updatedReplicas: 3, availableReplicas: 3, progress: 100 },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ deployments: fakeDeployments }),
-    })
-
-    const { result } = renderHook(() => useDeployments('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.deployments.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('tracks consecutive failures and returns lastRefresh', async () => {
-    // All fetch paths fail
-    mockIsAgentUnavailable.mockReturnValue(true)
-    globalThis.fetch = vi.fn().mockRejectedValue(new Error('Network error'))
-
-    const { result } = renderHook(() => useDeployments())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.consecutiveFailures).toBeGreaterThanOrEqual(1)
-    expect(result.current.lastRefresh).toBeDefined()
-  })
-})
-
-// ===========================================================================
-// useJobs
-// ===========================================================================
-
-describe('useJobs', () => {
-  it('returns initial loading state with empty jobs array', () => {
-    mockFetchSSE.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useJobs())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.jobs).toEqual([])
-  })
-
-  it('returns jobs after SSE fetch resolves', async () => {
-    const fakeJobs = [
-      { name: 'backup-job', namespace: 'system', cluster: 'c1', status: 'Complete', completions: '1/1', age: '1h' },
-    ]
-    mockFetchSSE.mockResolvedValue(fakeJobs)
-
-    const { result } = renderHook(() => useJobs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.jobs).toEqual(fakeJobs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns jobs from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeJobs = [
-      { name: 'migration-job', namespace: 'prod', status: 'Running', completions: '0/1' },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ jobs: fakeJobs }),
-    })
-
-    const { result } = renderHook(() => useJobs('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.jobs).toEqual(fakeJobs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('handles SSE failure with error message', async () => {
-    mockFetchSSE.mockRejectedValue(new Error('SSE error'))
-
-    const { result } = renderHook(() => useJobs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('SSE error')
-    expect(result.current.jobs).toEqual([])
-  })
-
-  it('provides refetch function', async () => {
-    mockFetchSSE.mockResolvedValue([])
-    const { result } = renderHook(() => useJobs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(typeof result.current.refetch).toBe('function')
-  })
-})
-
-// ===========================================================================
-// useHPAs
-// ===========================================================================
-
-describe('useHPAs', () => {
-  it('returns initial loading state with empty hpas array', () => {
-    mockFetchSSE.mockReturnValue(new Promise(() => {}))
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useHPAs())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.hpas).toEqual([])
-  })
-
-  it('returns HPAs from API after fetch resolves', async () => {
-    const fakeHPAs = [
-      { name: 'web-hpa', namespace: 'prod', reference: 'Deployment/web', minReplicas: 2, maxReplicas: 10, currentReplicas: 5 },
-    ]
-    mockApiGet.mockResolvedValue({ data: { hpas: fakeHPAs } })
-
-    const { result } = renderHook(() => useHPAs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.hpas).toEqual(fakeHPAs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns HPAs from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeHPAs = [
-      { name: 'api-hpa', namespace: 'prod', reference: 'Deployment/api', minReplicas: 1, maxReplicas: 5, currentReplicas: 3 },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ hpas: fakeHPAs }),
-    })
-
-    const { result } = renderHook(() => useHPAs('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.hpas).toEqual(fakeHPAs)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('API error'))
-
-    const { result } = renderHook(() => useHPAs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('API error')
-    expect(result.current.hpas).toEqual([])
-  })
-})
-
-// ===========================================================================
-// useReplicaSets
-// ===========================================================================
-
-describe('useReplicaSets', () => {
-  it('returns initial loading state with empty replicasets array', () => {
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useReplicaSets())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.replicaSets).toEqual([])
-  })
-
-  it('returns replicasets from API after fetch resolves', async () => {
-    const fakeRS = [
-      { name: 'web-rs-abc', namespace: 'prod', replicas: 3, readyReplicas: 3, ownerName: 'web', ownerKind: 'Deployment' },
-    ]
-    mockApiGet.mockResolvedValue({ data: { replicasets: fakeRS } })
-
-    const { result } = renderHook(() => useReplicaSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.replicaSets).toEqual(fakeRS)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns replicasets from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeRS = [
-      { name: 'api-rs-xyz', namespace: 'prod', replicas: 2, readyReplicas: 2 },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ replicasets: fakeRS }),
-    })
-
-    const { result } = renderHook(() => useReplicaSets('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.replicaSets).toEqual(fakeRS)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('API error'))
-
-    const { result } = renderHook(() => useReplicaSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('API error')
-  })
-})
-
-// ===========================================================================
-// useStatefulSets
-// ===========================================================================
-
-describe('useStatefulSets', () => {
-  it('returns initial loading state with empty statefulsets array', () => {
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useStatefulSets())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.statefulSets).toEqual([])
-  })
-
-  it('returns statefulsets from API after fetch resolves', async () => {
-    const fakeSS = [
-      { name: 'redis-0', namespace: 'data', replicas: 3, readyReplicas: 3, status: 'Running', image: 'redis:7' },
-    ]
-    mockApiGet.mockResolvedValue({ data: { statefulsets: fakeSS } })
-
-    const { result } = renderHook(() => useStatefulSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.statefulSets).toEqual(fakeSS)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns statefulsets from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeSS = [
-      { name: 'pg-0', namespace: 'data', replicas: 1, readyReplicas: 1, status: 'Running' },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ statefulsets: fakeSS }),
-    })
-
-    const { result } = renderHook(() => useStatefulSets('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.statefulSets).toEqual(fakeSS)
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('API error'))
-
-    const { result } = renderHook(() => useStatefulSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('API error')
-  })
-})
-
-// ===========================================================================
-// useDaemonSets
-// ===========================================================================
-
-describe('useDaemonSets', () => {
-  it('returns initial loading state with empty daemonsets array', () => {
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useDaemonSets())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.daemonSets).toEqual([])
-  })
-
-  it('returns daemonsets from API after fetch resolves', async () => {
-    const fakeDS = [
-      { name: 'node-exporter', namespace: 'monitoring', desiredScheduled: 3, currentScheduled: 3, ready: 3, status: 'Running' },
-    ]
-    mockApiGet.mockResolvedValue({ data: { daemonsets: fakeDS } })
-
-    const { result } = renderHook(() => useDaemonSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.daemonSets).toEqual(fakeDS)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns daemonsets from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeDS = [
-      { name: 'fluentd', namespace: 'logging', desiredScheduled: 5, currentScheduled: 5, ready: 5, status: 'Running' },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ daemonsets: fakeDS }),
-    })
-
-    const { result } = renderHook(() => useDaemonSets('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.daemonSets).toEqual(fakeDS)
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('API error'))
-
-    const { result } = renderHook(() => useDaemonSets())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('API error')
-  })
-})
-
-// ===========================================================================
-// useCronJobs
-// ===========================================================================
-
-describe('useCronJobs', () => {
-  it('returns initial loading state with empty cronjobs array', () => {
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => useCronJobs())
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.cronJobs).toEqual([])
-  })
-
-  it('returns cronjobs from API after fetch resolves', async () => {
-    const fakeCJ = [
-      { name: 'daily-backup', namespace: 'system', schedule: '0 2 * * *', suspend: false, active: 0 },
-    ]
-    mockApiGet.mockResolvedValue({ data: { cronjobs: fakeCJ } })
-
-    const { result } = renderHook(() => useCronJobs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.cronJobs).toEqual(fakeCJ)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns cronjobs from local agent when available', async () => {
-    mockIsAgentUnavailable.mockReturnValue(false)
-    const fakeCJ = [
-      { name: 'hourly-sync', namespace: 'ops', schedule: '0 * * * *', suspend: false, active: 1 },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ cronjobs: fakeCJ }),
-    })
-
-    const { result } = renderHook(() => useCronJobs('my-cluster'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.cronJobs).toEqual(fakeCJ)
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('API error'))
-
-    const { result } = renderHook(() => useCronJobs())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('API error')
-  })
-})
-
-// ===========================================================================
-// usePodLogs
-// ===========================================================================
-
-describe('usePodLogs', () => {
-  it('starts with empty logs and not loading (waits for params)', () => {
-    mockApiGet.mockReturnValue(new Promise(() => {}))
-    const { result } = renderHook(() => usePodLogs('c1', 'default', 'pod-1'))
-    // It sets loading to true then fetches
-    expect(result.current.logs).toBe('')
-  })
-
-  it('returns logs after API fetch resolves', async () => {
-    mockApiGet.mockResolvedValue({ data: { logs: 'line1\nline2\nline3' } })
-
-    const { result } = renderHook(() => usePodLogs('c1', 'default', 'pod-1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.logs).toBe('line1\nline2\nline3')
-    expect(result.current.error).toBeNull()
-  })
-
-  it('handles API failure with error message', async () => {
-    mockApiGet.mockRejectedValue(new Error('Not found'))
-
-    const { result } = renderHook(() => usePodLogs('c1', 'default', 'pod-1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.error).toBe('Not found')
-    expect(result.current.logs).toBe('')
-  })
-
-  it('passes container and tail params to the API', async () => {
-    mockApiGet.mockResolvedValue({ data: { logs: '' } })
-
-    const TAIL_LINES = 50
-    renderHook(() => usePodLogs('c1', 'default', 'pod-1', 'my-container', TAIL_LINES))
-
-    await waitFor(() => expect(mockApiGet).toHaveBeenCalled())
-    const url = mockApiGet.mock.calls[0][0] as string
-    expect(url).toContain('container=my-container')
-    expect(url).toContain('tail=50')
-  })
-
-  it('provides refetch function', async () => {
-    mockApiGet.mockResolvedValue({ data: { logs: 'log data' } })
-    const { result } = renderHook(() => usePodLogs('c1', 'default', 'pod-1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(typeof result.current.refetch).toBe('function')
-  })
-})
-
-// ===========================================================================
-// REGRESSION-PREVENTING ADDITIONS
-// ===========================================================================
-
-// ===========================================================================
-// useDeployments – extended coverage
-// ===========================================================================
-
